[Refactoring] Move ExpandSwitchCases and ExpandDefault to their own file
