+++ lib/Refactoring/CMakeLists.txt
@@ -5,6 +5,7 @@ add_swift_host_library(swiftRefactoring STATIC
   ConvertIfLetExprToGuardExpr.cpp
   ConvertToSwitchStmt.cpp
   ConvertToTernaryExpr.cpp
+  ExpandSwitchCases.cpp
   ExpandTernaryExpr.cpp
   ExtractExpr.cpp
   ExtractExprBase.cpp
